Make stable mir reachability default
